mockellobackend/backend/requirements.txt
fastapi
uvicorn
pymongo
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
mangum
requests
